+++ course/format/onetopic/lang/en/format_onetopic.php
@@ -86,3 +86,13 @@ $string['templatetopic_icons_help'] = 'This option defines if the icons are disp
 $string['hidden_message'] = 'The section <em>{$a}</em> is not currently available.';
 $string['privacy:metadata'] = 'The Onetopic format plugin does not store any personal data.';
 $string['hiddensectionshelp'] = 'Hidden sections are shown in collapsed form with available message';
+
+$string['tabsview'] = 'Tabs view';
+$string['tabsview_help'] = 'By default: is the traditional view.<br />
+Vertically: show tabs in vertical direction. Tabs on the left and content on the right.<br />
+One single line: all tabs are displayed in a single line with horizontal scroll. Useful if too many tabs are used.';
+$string['tabsview_default'] = 'By default';
+$string['tabsview_vertical'] = 'Vertically';
+$string['tabsview_oneline'] = 'Only one line';
+
+$string['increasesections'] = 'Add a section after the currently selected section';