+++ githash.php
@@ -23,4 +23,4 @@
  * and administration at http://docs.moodle.org/en/Git .
  */
 
-$githash = 'fa49207';
+$githash = '696857a';